+++ codex/views/opds/v1/entry/__init__.py
@@ -1,0 +1,1 @@
+"""OPDS v1 Entries."""
